requirements.txt
gdown
pyyaml
langchain
langchain_community
unstructured
python-docx


